:recycle: A field one width wider than both its ends reads as a truncation
`runtime/blit_exec`. `Type11Texture::row_stride` was `u64`. Both construction
sites fill it from `mapping_write::type11_sample_window` /
`type5_sample_window`, which return `u32`, and both readers hand it to
`SurfaceWindow::bpr`, which is `u32`. So every path was `surface_bpr as u64` in
and `t.row_stride as u32` out — a round trip, lossless, and indistinguishable at
the read site from an unchecked narrowing of a 64-bit guest field.

That is not hypothetical: an audit of narrowing casts on guest-decoded values
reported `bpr: t.row_stride as u32` as a live defect — "if guest provides
row_stride > 4GB, bpr silently truncates, reading wrong bytes per row" —
because the only thing visible at the cast is a `u64` field named after a
guest-decoded quantity. Settling it took reading both constructors. Now the
field is `u32` and there is nothing to settle; the two casts and the two
widenings are gone.

`LinearTextureLevel::row_stride` stays `u64` and is untouched. It is a different
field with a different provenance — `ld64` straight off the wire at
`TEXTURE_LEVEL_ROW_STRIDE` — and its readers do `checked_mul` in 64-bit, which
is what a 64-bit wire field needs.

No behavior change: the values that flowed through were `u32`-valued at both
ends before this and are the same after.

Also adjudicated in the same sweep, both false positives, neither changed:
`read_texture_row`'s `row_bytes as u32` and `(row_bytes / bpp) as u32` are
bounded by the `row_bytes as usize > buf.len()` refusal above them
(`rd_row_buf_cap`); and `fill_render_pso_key`'s `rt.slot as u8` on the Metal arm
has one call site, guarded 140 lines above by
`metal_render_color_slot_out_of_range` against
`REIMS_VGPU_METAL_MAX_COLOR_RTS = 8`.

Verified: 1611 passed, 0 failed, 6 ignored — unchanged, this adds no test
because it adds no behavior. Clippy `-D warnings` exit 0 on all three arms, one
per invocation. `cargo fmt --check` adds no file to the tree's standing list.

Not verified: no live boot; nothing here can change one.

+++ crates/reims-vgpu/src/runtime/blit_exec/mod.rs
@@ -374,7 +374,14 @@ struct Type11Texture {
     /// Byte offset of this texture/plane in the mapping allocation.
     surface_offset: u64,
     /// IOSurface-aligned surface row stride (bytes).
-    row_stride: u64,
+    ///
+    /// `u32` to match both ends it sits between: `type11_sample_window` and
+    /// `type5_sample_window` each return it as one, and its only readers hand
+    /// it to [`mapping_write::SurfaceWindow::bpr`], which is one. It was `u64`,
+    /// so both construction sites widened and both readers narrowed straight
+    /// back — a round trip that reads exactly like an unchecked truncation of a
+    /// 64-bit guest field and has been mistaken for one.
+    row_stride: u32,
     /// Exclusive end of the sample window (for page-span planning).
     span_end: u64,
     bpp: u32,
@@ -730,7 +737,7 @@ fn resolve_texture_backing_depth<M: HostMemory + HostOps>(
             width: tex_w,
             height: tex_h,
             surface_offset,
-            row_stride: surface_bpr as u64,
+            row_stride: surface_bpr,
             span_end,
             bpp,
             pixel_format: format,
@@ -817,7 +824,7 @@ fn resolve_texture_backing_depth<M: HostMemory + HostOps>(
             width: view.width,
             height: view.height,
             surface_offset,
-            row_stride: surface_bpr as u64,
+            row_stride: surface_bpr,
             span_end,
             bpp,
             pixel_format: format,
@@ -1014,7 +1021,7 @@ fn read_texture_row<M: HostMemory + HostOps>(
                 t.mapping_id,
                 mapping_write::SurfaceWindow {
                     base_off: t.surface_offset,
-                    bpr: t.row_stride as u32,
+                    bpr: t.row_stride,
                     span_end: t.span_end,
                     bpp: t.bpp,
                 },
@@ -1110,7 +1117,7 @@ fn write_texture_row<M: HostMemory + HostOps>(
                 t.mapping_id,
                 mapping_write::SurfaceWindow {
                     base_off: t.surface_offset,
-                    bpr: t.row_stride as u32,
+                    bpr: t.row_stride,
                     span_end: t.span_end,
                     bpp: t.bpp,
                 },
